CSS Modifications; Better Message Headers; Attempt at @NoCache; Avida Reconnections now resets the database; Messages page handles resets better
I made the CSS file make prettier formatting for the messages page.
However, there may still be some issues with array numbering starting at
zero.

nocache.py is supposed to provide header information for HTTP requests
that tells the browser not to pull from cache.  There might be an issue
with wsgi where the templates are still cached somewhere, so manual
refreshes are still necessary

.gitignore excludes some commonly generated python cache files as well
as any python notebooks I'm playing with.

I've modified the div class names to make things easier to understand
when working with the CSS.

Other misc changes such as displaying the update in the header of an
update av_debug message and preparing to handle av_ext_command messages.